Add some clarifying comments

+++ src/percentile_tracker.py
@@ -1,3 +1,5 @@
+from __future__ import division
+
 import bisect
 from collections import defaultdict, namedtuple
 import math
@@ -24,12 +26,12 @@ class PercentileTracker:
         
     def _nearest_rank_percentile(self, transaction_list):
         """Calculates the transaction value for the nearest"""
-        index = math.ceil(self.percentile_fraction*len(transaction_list)) - 1
+        index = int(math.ceil(self.percentile_fraction*len(transaction_list)) - 1)
         value = int(transaction_list[index] + .5)
         return value
         
     def is_repeat_donor(self, name, zip_code, year):
-        # Hash collision between two names in same zip code causes error.
+        # Hash collision between two names in same zip code causes inaccuracy.
         # Assume collision rare enough to not greatly affect results. 
         hashed_id = int(str(hash(name)) + zip_code)
         earliest_year = self.donor_dict.get(hashed_id, year + 1)
@@ -44,11 +46,13 @@ class PercentileTracker:
         
         :return stats: PercentileStats namedtuple instance
         """
+        # Get relevant transactions and update list
         recipient_hash = hash(recipient)
         transaction_list = self.recipient_dict[recipient_hash][int(zip_code)][year]
         insert_index = bisect.bisect(transaction_list, transaction_amt)
         transaction_list.insert(insert_index, transaction_amt)
 
+        # Calculate percentile stats for relevant transactions
         percentile_value = self._nearest_rank_percentile(transaction_list)
         transaction_sum = int(sum(transaction_list) + .5)
         stats = self.PercentileStats(
